Refactor RTC code into it's own subpackage
