+++ src/stactools/sentinel1/__init__.py
@@ -1,14 +1,15 @@
 import stactools.core
 
+from stactools.sentinel1.commands import create_sentinel1_command
+from stactools.cli.registry import Registry
+
 stactools.core.use_fsspec()
 
 
-def register_plugin(registry):
+def register_plugin(registry: Registry):
     # Register subcommands
 
-    from stactools.sentinel1 import commands
-
-    registry.register_subcommand(commands.create_sentinel1_command)
+    registry.register_subcommand(create_sentinel1_command)
 
 
 __version__ = '0.1.0'
