+++ src/stactools/sentinel1/rtc/__init__.py
@@ -1,0 +1,14 @@
+import stactools.core
+
+stactools.core.use_fsspec()
+
+
+def register_plugin(registry):
+    # Register subcommands
+
+    from stactools.sentinel1 import commands
+
+    registry.register_subcommand(commands.create_sentinel1_command)
+
+
+__version__ = '0.1.0'
