+++ src/stactools/sentinel1/rtc/constants.py
