+++ src/stactools/sentinel1/rtc/rtc_metadata.py
